Finished currency converter
Added get_currencies that takes api key and send a request for the currencies and organizes them into a list and sort them.

Added print_currencies that takes the cuurencies avaliable and print it out.

Added exchange rate and convert.  Exchange rate takes the symbol of the matching currencies and give the exchange rate between them.  "convert()" takes the two rates and return the conversion to the second currency.

Main() ties everything together and makes it a bit more easier to use